[#21] refactor : 통합테스트와 mvc테스트 분리
